Refactor config wizard
Deduplicated zig finder and added `zls config` subcommand
